Move common.js to /src
